krishirakshak-ml/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "krishirakshak-ml"
version = "0.1.0"
description = "ML subsystem for KrishiRakshak: detection/classification, export, AL, inference"
authors = [{name = "KrishiRakshak Team"}]
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1",
  "torchvision>=0.16",
  "torchaudio>=2.1; platform_system != 'Windows'",  # optional
  "ultralytics>=8.0.200",
  "timm>=0.9.8",
  "pydantic>=2.4",
  "python-dotenv>=1.0",
  "opencv-python-headless>=4.9",
  "numpy>=1.24",
  "pillow>=10.0",
  "scikit-learn>=1.3",
  "mlflow>=2.11",
  "fastapi>=0.110",
  "uvicorn>=0.23",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "rich>=13.7",
  "requests>=2.31",
  "scipy>=1.10",
  "matplotlib>=3.8",
  "seaborn>=0.13",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-cov>=4.1",
  "flake8>=6.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["krishirakshak-ml*"]

[tool.pytest.ini_options]
testpaths = ["krishirakshak-ml/tests"]
addopts = "-q"
